Fixes which may or may not break lots of stuff

+++ V3UPSManager/BackupChanges.cs
@@ -93,14 +93,22 @@ public partial class MainWindow : Form
 			Directory.CreateDirectory(directory);
 
 			// Back up the file by copying it to the desired backup folder
-			File.Copy(file, newpath);
+			if (!File.Exists(newpath))
+			{
+				Log("📋 VAR_BEFORE ⮞ VAR_AFTER", new Dictionary<string, string>()
+				{
+					{ "VAR_BEFORE", file },
+					{ "VAR_AFTER", newpath },
+				}, Verbosity.Info, LogType.ConsoleOnly);
+				File.Copy(file, newpath);
+			}
 		}
 	}
 
 	private void DisplayStatus()
 	{
 		// Count backup (containing "_bak") files in the installation folder
-		string[] files = Directory.GetFiles(installation_folder, "*.*_bak", SearchOption.AllDirectories);
+		string[] files = Directory.GetFiles(installation_folder, "*_bak", SearchOption.AllDirectories);
 		if (files != null && files.Length > 0 && files.Length == to_apply.Count)
 		{
 			Log(info[30]);
@@ -108,6 +116,17 @@ public partial class MainWindow : Form
 		else
 		{
 			Log(info[34], null, Verbosity.Error);
+			string apply_file = "to_apply.txt";
+			string bak_files = "bak_files.txt";
+			if (File.Exists(bak_files))
+			{
+				File.Delete(bak_files);
+			}
+			File.WriteAllText(bak_files, string.Join('\n', files));
+			if (!File.Exists(apply_file))
+			{
+				File.WriteAllText(apply_file, string.Join('\n', to_apply));
+			}
 		}
 	}
 }
